Multiple-value fields made explicit
As it turns out, Solr can only order by single-value fields - so, we
need to make an explicit distinction between single-value and
multiple-value fields in the configuration. Created separate
dynamicFields in Solr schema which are suffixed by 'm' to distinguish
them as multiple-value fields; modified field configurations in Sunspot
to take a :multiple => (true|false) parameter to determine which
fieldType is used to index the field.