Got rid of Builder API
I don't think it was all that intuitive and it certainly made the search
API fairly awkward. More to the point, I don't think it's a core part of
what Sunspot is trying to do. I will look into making it easy to apply
the Builder pattern (primarily by giving access to the search DSL
object) but I don't think Sunspot itself needs to provide a framework
for doing that.

I reimplemented standard support for passing in search parameters as a
hash, although only EqualTo, Between, and AnyOf restrictions can be
created in this way.