Refactor internals of Search/Query/Session
Search no longer delegates initialization of Query; rather, it just
receives a Query instance in its initializer, and provides access to the
query via a reader. Session takes care of building the query using the
provided block and/or applying the options hash.

Also implemented Session#new_search, which takes care of instantiating
the Search object without building it directly or executing it.