Refactor field architecture
Introduced FieldFactory concept; field factories are stored in Setup
objects based on input to the setup DSL. There is a FieldFactory for
static fields and one for dynamic fields; there is now only a single
Field type with reduced responsibilities. FieldFactories are responsible
for data extraction during the indexing process.