CompositeSetup refactor
Created CompositeSetup class, which holds a set of types and implements
a set of common methods - namely, #fields, #text_fields,
class as well where they weren't already. Setup/CompositeSetup are now
passed around in Query and Search objects and their appropriate
components to provide a unified interface/store for accessing fields by
name or as a collection for a given set of types.