Remove accidental ActiveSupport dependency
Sunspot accidentally depended on ActiveSupport in a couple of places
because the test suite loads WillPaginate, which loads ActiveSupport.
Refactored so that WillPaginate is only loaded (and the corresponding
tests only run) if an environment variable is passed; removed the
ActiveSupport dependencies.

Also, as part of this, made the Indexer object not specific to a
particular class/setup, since that required artificially doing more
indexing calls then necessary in certain cases and didn't really provide
any advantage.