Fix for connectives with nested negated restrictions
Lucene interprets negations in a somewhat non-intuitive way, which means
that naively combining a negated restriction with another restriction
using OR does not work as expected (in particular, all results must
match the negation, and at least one of the other restrictions in the
disjunction). Fix this by detecting negated restrictions inside
disjunctions and factoring out a NOT using DeMorgan's law.