Remove LocalSolr support
To date, I've been unable to find a way for LocalSolr to support all of
the query parameters Sunspot needs. It entirely ignores "fq" params,
which means normal filtering is out of the question. I also thought it
would be possible to put everything in the "q" param, but for keyword
search that requires a dismax subquery, and LocalSolr also isn't capable
of parsing subqueries.

Since there's no solution forthcoming, and I don't want to release
Sunspot with painfully incomplete support for local search, I'm pulling
it from v0.10. I'm pursuing a few possibilities for the future, but I
have no idea yet whether/when they'll pan out.