Complete refactor of faceting
Faceting concerns are now more separate, along the lines of the general
abstraction boundary between Sunspot::Query and the rest of the library.
In particular, Sunspot::Query is now only aware of faceting as it
directly applies to building queries - field facets, date facets, and
query facets, but nothing that maps to the higher-level facet
functionality (grouped query facets, :only field facets, etc.) that
Sunspot provides.

As well as helping to move further toward Sunspot::Query being a
standalone module without knowledge of the higher-level abstractions
elsewhere in the library, this change will make it easier to add further
faceting abstractions.