FieldFacet is a full subclass of QueryFacet
Search::FieldFacet is a specialized Search::QueryFacet. The reason for
this is that FieldFacet can often include QueryFacet rows (in some cases
those being the only ones, such as a limited field facet). So, a
FieldFacet is a QueryFacet that also looks for data in the facet_fields
response, and is capable of instantiating references.