Upgrade to trunk LocalSolr
Upgrade to trunk LocalSolr version. Biggest news here is that it allows
use of the :fq param, which means Sunspot finally can (and does) support
searches with both geo and fulltext components. Other news is that it's
compatible with Solr 1.4; it uses TrieFields, which are much more
efficient, for storage; and it switches from locallucene to
lucene-spatial, the latter being an official(-ish?) contrib in the Lucene
library.

This will, unfortunately, require reindexing of existing geographical data
(it's worth it).