Give DSL blocks access to calling context
When doing an instance_eval for a DSL block, now use a special class
that delegates to the block's calling context if the method is not
available on the DSL class itself. This means that, assuming no
namespace conflict, users can refer to instance method in the calling
context even though the block is being instance_eval'd in the DSL class.

Thanks much to Dmitry Dzema for the idea:
http://blog.dzema.name/2009/11/ruby-dsl-and-instance_eval