Added --force option to installer; use it on packaged schema
Added a --force option to the installer, which overwrites Sunspot's
fields and types with Sunspot's preferred configuration. This is needed
when bootstrapping a new packaged Solr schema, since the built-in schema
has certain naming conventions and type decisions which break Sunspot if
left intact. Applied the installer with --force to the default schema;
this is now Sunspot's shipped schema. For some reason, geo+fulltext
search is now broken - haven't figured out why yet.