Fix memory leak in text field setups with Rails class reloading
The use of a hash of arrays for the @stored_field_factories_cache was
causing a memory leak in cases where class reloading was used, because
the array was being appended to each time the setup block was run.
Changed it so that the stored field for a given hit and given field name
is the last stored field with that name configured for that class. This
should fix the memory leak and also seems a bit cleaner to me.