Reinitialize Setup objects if class's object ID has changed
This basically bulletproofs Sunspot setups against class reloading
(like in development mode in Rails). As well as a recent memory leak
that came up, this will eliminate a rare but confusing situation in
which a field is removed from a setup in development mode, the
application isn't restarted, and Sunspot still expects the field to be
there.