Properly namespace thread-local session instances
Previously I was storing the session wrapped by a
ThreadLocalSessionProxy in a static thread-local session variable,
:sunspot_session. This of course meant that two ThreadLocalSessionProxy
instances would share the same underlying Session instance in a given
thread, which is not the expected or desired behavior (e.g., when two
TLSPs are encapsulated by a MasterSlaveSessionProxy).

Now the underlying session is stored in a thread-local variable which
includes the proxy's object_id, thus scoping each session to its owning
proxy. A finalizer is used to ensure that the session instances don't
remain in the thread-local namespace after the proxy instance is garbage
collected (a concern if threads are long-lived but proxies are
short-lived; granted an unlikely use case in reality, but due diligence
never hurts).

[#103 state:resolved]