Hit#stored_value shouldn't treat nil and false the same way.  Fixed.
As far as I can tell, Sunspot uses nil to represent the non-existence of
a value for indexing or (in thie case) data stored in the Solr index.
False is a valid value for stored data of boolean type; however, the
code in Hit#stored_value prior to this commit did not differentiate
between the two.

This commit contains a change intended to remedy that problem.