No need to escape arguments if passed explicitly
The use of Shellwords caused the sunspot-solr stop command to stop
working, because the exec call first created an intermediate system
shell, to handle escaped parameters. The registered pid is that of the
shell, terminating the shell does not terminate the java command.

Instead of trying to escape arguments it's possible to pass in arguments
explicitly to exec (and related calls such as system), this avoids the
possibility of an intermediate shell and any problems with escaping

(cherry picked from commit 695aa9cfac74fd096f4b605ccb7ebfa4ed150bc9)