Add CompositeSetup#more_like_this_fields
Sunspot#more_like_this searches were failing if multiple types were provided
to run the search over. This is because a Sunspot::CompositeSetup object is
used to extract the fields in this case, instead of the usual
Sunspot::Setup object. The CompositeSetup class was missing the required
more_like_this_fields method, which is added by this commit.

A test is also added to the more_like_this integration tests.