Allow query parser to be optionally set to dismax
This commit adds the option to set the query parser to dismax via a
user_configuration variable `query_parser` in the sunspot.yml. If the
option is not used then edismax is set as the default.

The reason for this is to help support those who need to interface older
Solr setups w/o having to monkey patch.