Improve joins when using in filter queries so they join columns from the desired models only. (#956)
This fixes the case when a few models have a column with the same name, e.g. "published". For example, a user has a Profile and Blog, both have published:boolean, but you need to filter users with a published Profile only.

The current implementation will fail and return users with both published Profile and Blog, since both have published_b and member_id_i columns in the index.

Additionally, allow join targets to be defined as strings to prevent circular dependency errors