Ability to explicitly provide ID prefix when deleting by ID.
Sometimes ID prefix could be known and explicitly provided to the delete_by_id in this format "USERDATA!User".

Example use-case – indexing queues, when synchronization with Solr happens in background. Record being removed from the database and a corresponding event created in the queue with record_class_name, record_id, type of the operation, and ID prefix, which is calculated before removing the record.